did requested changes
