+++ src/main/java/util/InsufficientInventoryException.java
@@ -1,4 +1,4 @@
-package service;
+package util;
 
 public class InsufficientInventoryException extends RuntimeException {
     public InsufficientInventoryException(String message) {
