+++ src/main/java/util/StrategyException.java
@@ -1,4 +1,4 @@
-package service;
+package util;
 
 public class StrategyException extends RuntimeException {
     public StrategyException(String message) {
