+++ src/main/java/util/ValidationDataException.java
@@ -1,4 +1,4 @@
-package service;
+package util;
 
 public class ValidationDataException extends RuntimeException {
     public ValidationDataException(String message) {
